python-scripts/requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0 